Fix cast issues?

+++ embedded/devices/src/ina3221.cpp
@@ -112,7 +112,7 @@ Reads bus voltage for a specified channel. Returns volts.
 float Ina3221::readVoltage(int channel) {
   uint16_t voltage_from_reg;
   int reg;
-  float voltage;
+  int voltage;
   int rc;
 
   if (channel < 1 || channel > INA_NUM_CHANNELS) {
@@ -125,12 +125,12 @@ float Ina3221::readVoltage(int channel) {
   if (rc) return rc;
 
   // Get actual value from register
-  voltage = (float)((voltage_from_reg >> 3) & INA_VOLTAGE_MASK);
+  voltage = (int)((voltage_from_reg >> 3) & INA_VOLTAGE_MASK);
 
   // Preserve signedness
   if (voltage_from_reg & INA_VOLTAGE_SIGN) voltage *= -1;
 
-  return voltage * INA_BVOLTAGE_MULT;
+  return (float)voltage * INA_BVOLTAGE_MULT;
 }
 
 /*
@@ -139,7 +139,7 @@ Reads current for a specific channel. Returns amps.
 float Ina3221::readCurrent(int channel) {
   uint16_t voltage_from_reg;
   int reg;
-  float voltage;
+  int voltage;
   int rc;
 
   if (channel < 1 || channel > INA_NUM_CHANNELS) {
@@ -152,11 +152,11 @@ float Ina3221::readCurrent(int channel) {
   if (rc) return rc;
 
   // Get actual value from register
-  voltage = (float)((voltage_from_reg >> 3) & INA_VOLTAGE_MASK);
+  voltage = (int)((voltage_from_reg >> 3) & INA_VOLTAGE_MASK);
 
   // Preserve signedness
   if (voltage_from_reg & INA_VOLTAGE_SIGN) voltage *= -1;
 
   // Return voltage divided by shunt resistor
-  return (voltage * INA_SVOLTAGE_MULT) / this->shunts[channel - 1];
+  return ((float)voltage * INA_SVOLTAGE_MULT) / this->shunts[channel - 1];
 }
